Huge fix for wallet related interfaces
